requirements-test.txt
# Azure OpenAI Integration Testing Dependencies

# Core OpenAI library with Azure support
openai>=1.3.0,<2.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# JSON handling and utilities
pydantic>=2.0.0,<3.0.0

# HTTP client with better error handling
httpx>=0.24.0,<1.0.0

# Async support (if needed for future tests)
aiohttp>=3.8.0,<4.0.0

# Testing utilities
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.10.0,<4.0.0

# Logging and debugging
colorlog>=6.7.0,<7.0.0

# Development dependencies
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0

langchain-openai
python-dotenv